pyproject.toml
[tool.poetry]
name = "raptor-xp"
version = "1.0.0"
description = "View Raptor XP"
authors = ["vcokltfre <vcokltfre@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
aiomysql = "^0.1.0"
disnake = "^2.4.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[tool.pyright]
reportMissingTypeStubs = false
reportUnknownVariableType = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
